+++ src/api/estimate-coin-sell-all.js
@@ -1,26 +1,28 @@
 import {stringify as qsStringify} from 'qs';
 import {convertFromPip, convertToPip} from 'minterjs-util';
 // import {convertFromPip, convertToPip} from 'minterjs-util/src/converter.js';
-import {isCoinId, isValidNumber} from '../utils.js';
+import {isValidNumber, isCoinId} from '../utils.js';
 
 /**
  * @typedef {Object} EstimateSellAllResult
  * @property {number|string} will_get - amount of coinToBuy
- * @property {"pool"|"bancor"} swap_from
+ * @property {"pool"|"bancor"} [swap_from]
  */
 
 /**
  * @param {MinterApiInstance} apiInstance
- * @return {function({coinToSell: (string|number), coinToBuy: (string|number), swapFrom?: ESTIMATE_SWAP_TYPE, route?: Array<string|number>}, axiosOptions: AxiosRequestConfig=): Promise<EstimateSellAllResult>}
+ * @return {function({coinToSell: string, valueToSell: (string|number), coinToBuy: string, swapFrom?: ESTIMATE_SWAP_TYPE, route?: Array<string|number>, gasPrice: (string|number)}, axiosOptions: AxiosRequestConfig=): Promise<EstimateSellAllResult>}
  */
 export default function EstimateCoinSellAll(apiInstance) {
     return estimateCoinSellAll;
     /**
      * @param {Object} params
      * @param {string|number} params.coinToSell - ID or symbol of the coin to sell
+     * @param {string|number} params.valueToSell
      * @param {string|number} params.coinToBuy - ID or symbol of the coin to buy
      * @param {ESTIMATE_SWAP_TYPE} [params.swapFrom] - estimate pool swap
      * @param {Array<string|number>} [params.route] - intermediate coins IDs for pool swaps
+     * @param {string|number} [params.gasPrice]
      * @param {import('axios').AxiosRequestConfig} [axiosOptions]
      * @return {Promise<EstimateSellAllResult>}
      */
@@ -39,6 +41,9 @@ export default function EstimateCoinSellAll(apiInstance) {
         if (!params.coinToBuy && params.coinToBuy !== 0) {
             return Promise.reject(new Error('Coin to buy not specified'));
         }
+        if (!params.valueToSell) {
+            return Promise.reject(new Error('Value to sell not specified'));
+        }
         if (!params.coinToSell && params.coinToSell !== 0) {
             return Promise.reject(new Error('Coin to sell not specified'));
         }
@@ -46,13 +51,14 @@ export default function EstimateCoinSellAll(apiInstance) {
         params = {
             coin_id_to_sell: isCoinId(params.coinToSell) ? params.coinToSell : undefined,
             coin_to_sell: !isCoinId(params.coinToSell) ? params.coinToSell : undefined,
+            value_to_sell: convertToPip(params.valueToSell),
             coin_id_to_buy: isCoinId(params.coinToBuy) ? params.coinToBuy : undefined,
             coin_to_buy: !isCoinId(params.coinToBuy) ? params.coinToBuy : undefined,
             swap_from: params.swapFrom,
             route: params.route,
+            gas_price: params.gasPrice,
         };
 
-        // @TODO incorrect path
         return apiInstance.get('estimate_coin_sell_all', {
             ...axiosOptions,
             params,
@@ -63,15 +69,11 @@ export default function EstimateCoinSellAll(apiInstance) {
                 if (!isValidNumber(resData.will_get)) {
                     throw new Error('Invalid estimation data, `will_get` not specified');
                 }
-                if (!isValidNumber(resData.commission)) {
-                    throw new Error('Invalid estimation data, `commission` not specified');
-                }
 
                 return {
                     ...resData,
                     // receive pips from node and convert them
                     will_get: convertFromPip(resData.will_get),
-                    commission: convertFromPip(resData.commission),
                 };
             });
     }